:truck: 패키지 구조 변경
- api별 패키지 구조 변경
