+++ noriter/src/main/java/com/codewarts/noriter/article/domain/Image.java
@@ -1,0 +1,29 @@
+package com.codewarts.noriter.article.domain;
+
+import com.codewarts.noriter.article.domain.Article;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Image {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn
+  private Article article;
+
+  private String imageUrl;
+  private boolean deleted;
+
+}
