+++ noriter/src/main/java/com/codewarts/noriter/article/domain/Question.java
@@ -1,0 +1,14 @@
+package com.codewarts.noriter.article.domain;
+
+import com.codewarts.noriter.article.domain.Article;
+import javax.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Question extends Article {
+
+  private boolean solved;
+
+}
