+++ noriter/src/main/java/com/codewarts/noriter/article/domain/type/ArticleType.java
@@ -1,0 +1,5 @@
+package com.codewarts.noriter.article.domain.type;
+
+public enum ArticleType {
+  QUESTION, FREE, STUDY
+}
