+++ noriter/src/main/java/com/codewarts/noriter/auth/jwt/JwtProperties.java
@@ -1,0 +1,17 @@
+package com.codewarts.noriter.auth.jwt;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
+
+@Getter
+@RequiredArgsConstructor
+@ConstructorBinding
+@ConfigurationProperties(prefix = "user-jwt")
+public class JwtProperties {
+
+    private final String issuer;
+    private final String secretKey;
+
+}
