+++ noriter/src/main/java/com/codewarts/noriter/auth/oauth/dto/GithubAccessTokenResponse.java
@@ -1,0 +1,20 @@
+package com.codewarts.noriter.auth.oauth.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Getter;
+
+@Getter
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class GithubAccessTokenResponse {
+
+    private String accessToken;
+    private String tokenType;
+
+    public OAuthAccessToken toOAuthAccessToken() {
+        return OAuthAccessToken.builder()
+            .tokenValue(accessToken)
+            .tokenType(tokenType)
+            .build();
+    }
+}
