+++ noriter/src/main/java/com/codewarts/noriter/auth/oauth/properties/OAuthProperties.java
@@ -1,0 +1,16 @@
+package com.codewarts.noriter.auth.oauth.properties;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class OAuthProperties {
+
+    private final String callbackUrl;
+    private final String clientId;
+    private final String clientSecret;
+    private final String accessTokenApiUrl;
+    private final String loginFormUrl;
+    private final String userInfoApiUrl;
+}
