+++ noriter/src/main/java/com/codewarts/noriter/auth/oauth/type/ResourceServer.java
@@ -1,0 +1,5 @@
+package com.codewarts.noriter.auth.oauth.type;
+
+public enum ResourceServer {
+    GITHUB
+}
