+++ noriter/src/main/java/com/codewarts/noriter/auth/utils/OAuthUtils.java
@@ -1,0 +1,8 @@
+package com.codewarts.noriter.auth.utils;
+
+public class OAuthUtils {
+
+    public static final String ACCESS_TOKEN = "Access-Token";
+    public static final String REFRESH_TOKEN = "Refresh-Token";
+
+}
