+++ noriter/src/main/java/com/codewarts/noriter/common/config/WebConfig.java
@@ -1,0 +1,28 @@
+package com.codewarts.noriter.common.config;
+
+
+import io.netty.resolver.DefaultAddressResolverGroup;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.reactive.ReactorClientHttpConnector;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.netty.http.client.HttpClient;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig {
+
+    @Bean
+    public HttpClient httpClient() {
+        return HttpClient.create()
+            .resolver(DefaultAddressResolverGroup.INSTANCE);
+    }
+
+    @Bean
+    public WebClient webClient(HttpClient httpClient) {
+        return WebClient.builder()
+            .clientConnector(new ReactorClientHttpConnector(httpClient))
+            .build();
+    }
+}
