+++ noriter/src/main/java/com/codewarts/noriter/common/repository/MemberRepository.java
@@ -1,0 +1,13 @@
+package com.codewarts.noriter.common.repository;
+
+import com.codewarts.noriter.common.domain.Member;
+import com.codewarts.noriter.auth.oauth.type.ResourceServer;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByResourceServerAndResourceServerId(ResourceServer resourceServer, Long resourceServerId);
+}
