--- noriter/src/main/java/com/codewarts/noriter/domain/article/Study.java
@@ -1,13 +1,0 @@
-package com.codewarts.noriter.domain.article;
-
-import javax.persistence.Entity;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Study extends Article {
-
-  private boolean completed;
-
-}
