--- noriter/src/main/java/com/codewarts/noriter/domain/type/ArticleType.java
@@ -1,5 +1,0 @@
-package com.codewarts.noriter.domain.type;
-
-public enum ArticleType {
-  QUESTION, FREE, STUDY
-}
