--- noriter/src/main/java/com/codewarts/noriter/oauth/OAuthService.java
@@ -1,10 +1,0 @@
-package com.codewarts.noriter.oauth;
-
-import com.codewarts.noriter.domain.Member;
-
-public interface OAuthService {
-
-    OAuthAccessToken reqeustAccessToken(String code);
-
-    Member reqeustUserInfo(OAuthAccessToken oauthAccessToken);
-}
