--- noriter/src/main/java/com/codewarts/noriter/oauth/ResourceServer.java
@@ -1,5 +1,0 @@
-package com.codewarts.noriter.oauth;
-
-public enum ResourceServer {
-    GITHUB
-}
