--- noriter/src/main/java/com/codewarts/noriter/oauth/jwt/JwtProperties.java
@@ -1,17 +1,0 @@
-package com.codewarts.noriter.oauth.jwt;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
-
-@Getter
-@RequiredArgsConstructor
-@ConstructorBinding
-@ConfigurationProperties(prefix = "user-jwt")
-public class JwtProperties {
-
-    private final String issuer;
-    private final String secretKey;
-
-}
