--- noriter/src/main/java/com/codewarts/noriter/oauth/utils/OAuthUtils.java
@@ -1,8 +1,0 @@
-package com.codewarts.noriter.oauth.utils;
-
-public class OAuthUtils {
-
-    public static final String ACCESS_TOKEN = "Access-Token";
-    public static final String REFRESH_TOKEN = "Refresh-Token";
-
-}
